+++ src/main/java/auth/login/LoginMember.java
@@ -1,4 +1,4 @@
-package auth;
+package auth.login;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
