+++ src/main/java/auth/token/TokenResponse.java
@@ -1,4 +1,4 @@
-package auth;
+package auth.token;
 
 public class TokenResponse {
     private String accessToken;
