+++ src/main/java/nextstep/waiting/WaitingController.java
@@ -1,6 +1,6 @@
 package nextstep.waiting;
 
-import auth.LoginMember;
+import auth.login.LoginMember;
 import auth.UserDetails;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
