+++ src/main/java/nextstep/waiting/WaitingRequest.java
@@ -3,7 +3,8 @@ package nextstep.waiting;
 public class WaitingRequest {
     private Long scheduleId;
 
-    public WaitingRequest() {}
+    public WaitingRequest() {
+    }
 
     public WaitingRequest(Long scheduleId) {
         this.scheduleId = scheduleId;
