+++ src/test/java/nextstep/theme/ThemeE2ETest.java
@@ -1,7 +1,7 @@
 package nextstep.theme;
 
-import auth.TokenRequest;
-import auth.TokenResponse;
+import auth.token.TokenRequest;
+import auth.token.TokenResponse;
 import io.restassured.RestAssured;
 import nextstep.AbstractE2ETest;
 import nextstep.member.MemberRequest;
